+++ lib/src/http/session/session.dart
@@ -8,8 +8,7 @@ import 'package:meta/meta.dart';
 import 'package:crypto/crypto.dart';
 
 import 'package:jaguar/src/http/context/context.dart';
-import 'package:jaguar/src/http/request/request.dart';
-import 'package:jaguar/src/http/response/import.dart';
+import 'package:jaguar/src/http/response/response.dart';
 
 part 'session_on_cookie.dart';